04/19/2022 Tue

+++ utils/utils.py
@@ -7,6 +7,7 @@ import win32com.client
 import re
 import threading
 import time
+import glob
 import itertools
 from difflib import SequenceMatcher
 import os
@@ -42,46 +43,134 @@ class DB_Utils():
 
     def insert_dataframe(self, tablename, df):
         cols = ', '.join(str(x) for x in df.columns)
-        print(tablename, cols)
+        print('[EVENT] TABLE NAME: ', tablename, cols)
         for i in range(len(df)):
             time.sleep(0.02)
             sql = f'''INSERT INTO {tablename} ({cols}) VALUES {tuple(df[df.columns].values[i])}'''
             self.cursor.execute(sql)
         self.conn.commit() 
 
+import pandas as pd
+import numpy as np
+import pyodbc
+from datetime import datetime
+import win32com
+import win32com.client
+import re
+import threading
+import time
+import itertools
+from difflib import SequenceMatcher
+import os
+
+
+class Global(object):
+    root_path = 'C:/Users/qcells/Desktop/ATP/US_ATP_BI'
+
+class DB_Utils():
+    def connect_sapdb(self):
+        pass
+
+    def connect_azuredb(self):
+        server = 'qcells-us-atp-db-server.database.windows.net'
+        database = 'us-qcells-atp-db'
+        username = 'qcells'
+        password = '{Asdqwe123!@#}'   
+        driver= '{ODBC Driver 17 for SQL Server}'  
+        self.conn = pyodbc.connect('DRIVER='+driver
+                                +';SERVER=tcp:'+server
+                                +';PORT=1433;DATABASE='+database
+                                +';UID='+username
+                                +';Pwd='+password
+                                +';Encrypt=yes;TrustServerCertificate=no;Connection Timeout=30;')
+        self.cursor = self.conn.cursor()
+    
+    def fetch_data(self, sql):
+        self.cursor.execute(sql)
+        row = self.cursor.fetchall()
+        row = [list(i) for i in row]
+        col_names = [item[0] for item in self.cursor.description]
+        return pd.DataFrame(row, columns=col_names)
+
+    def insert_dataframe(self, tablename, df):
+        cols = ', '.join(str(x) for x in df.columns)
+        print(tablename, cols)
+        for i in range(len(df)):
+            try:
+                time.sleep(0.01)
+                sql = f'''INSERT INTO {tablename} ({cols}) VALUES {tuple(df[df.columns].values[i])}'''
+                self.cursor.execute(sql)
+            except Exception as e:
+                print(e)
+                print(sql)
+                pass
+        self.conn.commit() 
+
 class ETL_Utils(DB_Utils):
-    def meta_sheet(self, excel_name):
-        sql = f'''select sheet_list, primary_key, target_columns from meta_sheet_info
-                where file_name = {excel_name}'''
+    def __init__(self, excel_name):
+        self.excel_name = excel_name
+        self.connect_azuredb()
+        self.path = '../data/dummy/{}.xlsx'.format(self.excel_name)
+
+    def meta_sheet_info(self):
+        sql = f'''select sheetlist, targetcolumns, datatype from SHEET_INFO
+                where ExcelName = '{self.excel_name}' and UseYN = 'Y'
+                '''
         self.sheet_info = self.fetch_data(sql)
+        self.target_sheet_list = list(set(self.sheet_info['sheetlist'].tolist()))
 
-    def clean_data(self):
-        pk = self.sheet_info[self.sheet_info == sheet]['primary_key'].values[0]
-        target_cols = self.sheet_info[self.sheet_info == sheet]['target_columns'].tolist()
+    def read_excel(self):
+        if self.path[-4:] == 'xlsb':
+            df_obj = pd.ExcelFile(self.path,engine='pyxlsb')
+        else:
+            df_obj = pd.ExcelFile(self.path,engine='openpyxl')
         
-        df = df.dropna(subset = pk).reset_index(drop = True)
-        df = df[target_cols]
-        return df
-
-    def read_excel(self, path):
+        df = df_obj.parse(sheet_name=self.target_sheet_list[0])
+        df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+        self.df = df.dropna(how = 'all')
+    
+    def data_text(self, x):
+        return str(x)
+    
+    def yyyymmdd_datetime(self, x):
+        removal_string = ['/', '.', '-', '_', ' ']    
+        for rs in removal_string:
+            x = x.replace(rs, '')
+        if len(x) == 6:
+            x = x + '01'
         try:
-            if path[-4:] == 'xlsb':
-                df_obj = pd.ExcelFile(path,engine='pyxlsb')
-            else:
-                df_obj = pd.ExcelFile(path,engine='openpyxl')
-
-            self.df_list = []
-            self.meta_sheet(path)
-
-            for sheet in self.sheet_info['sheet_list'].tolist():
-                df = df_obj.parse(sheet_name=sheet)
-                df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
-                df = self.clean_data(df)
-                self.df_list.append(df)
-        except Exception as e:
-            print(e)
-            self.send_email(path, e)
+            return datetime.strftime(datetime.strptime(x, '%Y%m%d'), '%Y-%m-%d')
+        except:
+            return datetime.strftime(datetime.strptime(x, '%m%d%Y'), '%Y-%m-%d')
             pass
+
+    def clean_data_text(self):
+        target_columns_text = self.sheet_info[self.sheet_info['datatype'] == 'TEXT']['targetcolumns'].tolist()
+        print(target_columns_text, '\n')
+        for col in target_columns_text:
+            self.df[col] = self.df[col].apply(lambda x:str(x))
+
+    def clean_data_int(self):
+        target_columns_int = self.sheet_info[self.sheet_info['datatype'] == 'INT']['targetcolumns'].tolist()
+        print(target_columns_int, '\n')
+        for col in target_columns_int:
+            self.df[col] = self.df[col].apply(lambda x:int(x))
+
+    def clean_data_float(self):
+        target_columns_float = self.sheet_info[self.sheet_info['datatype'] == 'FLOAT']['targetcolumns'].tolist()
+        print(target_columns_float, '\n')
+        for col in target_columns_float:
+            self.df[col] = self.df[col].apply(lambda x:float(x))
+
+    def clean_data_datetime(self):
+        target_columns_datetime = self.sheet_info[self.sheet_info['datatype'] == 'DATETIME']['targetcolumns'].tolist()
+        print(target_columns_datetime, '\n')
+
+        for col in target_columns_datetime:
+            self.df[col] = self.df[col].apply(lambda x:self.yyyymmdd_datetime(str(x)))
+            
+        self.insert_dataframe(self.excel_name, self.df)
+        self.conn.close()
 
 class SAP_Item_Code(DB_Utils):
     def __init__(self):
@@ -108,16 +197,19 @@ class Master_Reset(DB_Utils):
 
     def reset_master_tables(self):
         for sheet in self.master_df.sheet_names:
-            sql = 'DELETE FROM {}'.format(sheet)
-            self.cursor.execute(sql)
+            if sheet != 'US_CITY_CODE':
+                sql = 'DELETE FROM {}'.format(sheet)
+                self.cursor.execute(sql)
         self.conn.commit()
 
-    def push_data(self):
+    def push_data(self, file_path = '/data/master.xlsx'):
         print('[EVENT] NEW MASTER DATA IS BEING SAVED !')
         for sheet_name in self.master_df.sheet_names:
-            df = pd.ExcelFile(Global.root_path + '/data/master.xlsx',engine='openpyxl').parse(sheet_name)
-            df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
-            self.insert_dataframe(sheet_name, df)
+            if sheet_name != 'US_CITY_CODE':
+                df = pd.ExcelFile(Global.root_path + file_path,engine='openpyxl').parse(sheet_name)
+                df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+                df.dropna(inplace = True, how = 'all')
+                self.insert_dataframe(sheet_name, df)
         print('\n')
         
     def master_reset_main(self, sender_addr): 
@@ -129,15 +221,22 @@ class Master_Reset(DB_Utils):
             self.master_df = pd.ExcelFile(Global.root_path + '/data/master.xlsx',engine='openpyxl')
             self.master_df.sheet_names #equal to DB table name     
             self.reset_master_tables()
-            self.push_data()
+            
+            try:
+                self.push_data()
+            except Exception as e:
+                print('[WARNING] MASTER FILE IS DAMAGED. RPA WILL ROLL-BACK.')
+                print(e)
+                self.reset_master_tables()
+                self.push_data(file_path = glob.glob('/data/MASTER_HIST/*.xlsx')[-1])
+                pass
+
             self.conn.close()
             return True
         else:
             print('[WARNING] send e-mail that rejected due to low-authorization')
             self.conn.close()
             return False
-        
-
         
 class Email_Utils(Master_Reset):
     def __init__(self, mail_receivers):
@@ -230,10 +329,10 @@ class Email_Utils(Master_Reset):
                             for att in atts:
                                 if att.FileName == 'master.xlsx':
                                     os.makedirs(Global.root_path + '/data/MASTSER_HIST', exist_ok = True)
-                                    att.SaveAsFile(Global.root_path + '/data/MASTSER_HIST/' + i.SentOn.strftime('%Y%m%d%H%M%S') + '_' + att.FileName) #saving Backup file
                                     att.SaveAsFile(Global.root_path + '/data/' + att.FileName) # saving Master file
                                     reset_result = self.master_reset_main(sender_addr)
                                     if reset_result == True:
+                                        att.SaveAsFile(Global.root_path + '/data/MASTSER_HIST/' + i.SentOn.strftime('%Y%m%d%H%M%S') + '_' + att.FileName) #saving Backup file
                                         self.send_email('[RPA] MASTER FILE RESET RESULT', 'This is the file used!'
                                                         ,destination = self.sender_mailaddr_extract(i)
                                                         ,master = True)
